checkpoint: finish restructuring the lib folder
