+++ lib/presentation/my_health/pages/my_health_page.dart
@@ -7,6 +7,7 @@ import 'package:afya_moja_core/information_list_card.dart';
 import 'package:async_redux/async_redux.dart';
 import 'package:flutter_svg/svg.dart';
 import 'package:misc_utilities/misc.dart';
+import 'package:myafyahub/presentation/profile/health_timeline/my_health_timeline.dart';
 import 'package:shared_themes/spaces.dart';
 import 'package:shared_themes/text_themes.dart';
 
@@ -21,7 +22,6 @@ import 'package:myafyahub/presentation/core/theme/theme.dart';
 import 'package:myafyahub/presentation/core/widgets/app_bar/custom_app_bar.dart';
 import 'package:myafyahub/presentation/core/widgets/custom_scaffold/app_scaffold.dart';
 import 'package:myafyahub/presentation/my_health/widgets/profile_health_details_widget.dart';
-import 'package:myafyahub/presentation/profile/widgets/my_health_timeline.dart';
 import 'package:myafyahub/presentation/router/routes.dart';
 
 class MyHealthPage extends StatelessWidget {
