+++ lib/presentation/profile/clinic_information/clinic_information_page.dart
@@ -10,7 +10,7 @@ import 'package:myafyahub/domain/core/value_objects/app_strings.dart';
 import 'package:myafyahub/presentation/core/theme/theme.dart';
 import 'package:myafyahub/presentation/core/widgets/app_bar/custom_app_bar.dart';
 import 'package:myafyahub/presentation/profile/widgets/call_contact_action_widget.dart';
-import 'package:myafyahub/presentation/profile/widgets/clinic_information_item_widget.dart';
+import 'package:myafyahub/presentation/profile/clinic_information/clinic_information_item_widget.dart';
 
 /// [ClinicInformationPage] is to display information of the clinic a client is registered to
 ///
