+++ lib/presentation/profile/faqs/profile_faqs_page.dart
@@ -10,7 +10,7 @@ import 'package:shared_themes/text_themes.dart';
 import 'package:myafyahub/domain/core/entities/faqs/faq_content.dart';
 import 'package:myafyahub/domain/core/value_objects/app_strings.dart';
 import 'package:myafyahub/presentation/core/widgets/app_bar/custom_app_bar.dart';
-import 'package:myafyahub/presentation/profile/widgets/faq_item.dart';
+import 'package:myafyahub/presentation/profile/faqs/faq_item.dart';
 import 'package:myafyahub/presentation/router/routes.dart';
 
 class ProfileFaqsPage extends StatelessWidget {
