+++ lib/presentation/profile/feedback/feedback_page.dart
@@ -14,7 +14,7 @@ import 'package:myafyahub/domain/core/value_objects/app_widget_keys.dart';
 import 'package:myafyahub/domain/core/value_objects/enums.dart';
 import 'package:myafyahub/presentation/core/theme/theme.dart';
 import 'package:myafyahub/presentation/core/widgets/app_bar/custom_app_bar.dart';
-import 'package:myafyahub/presentation/profile/widgets/medical_data_item_title.dart';
+import 'package:myafyahub/presentation/profile/medical_data/medical_data_item_title.dart';
 
 // Package imports
