+++ lib/presentation/profile/health_timeline/my_health_timeline.dart
@@ -13,7 +13,7 @@ import 'package:myafyahub/domain/core/value_objects/app_strings.dart';
 import 'package:myafyahub/presentation/core/theme/theme.dart';
 import 'package:myafyahub/presentation/profile/widgets/custom_timeline_list_item.dart';
 import 'package:myafyahub/presentation/profile/widgets/dashed_line.dart';
-import 'package:myafyahub/presentation/profile/widgets/timeline_indicator.dart';
+import 'package:myafyahub/presentation/profile/health_timeline/timeline_indicator.dart';
 
 class MyHealthTimeline extends StatefulWidget {
   const MyHealthTimeline({Key? key}) : super(key: key);
