--- lib/presentation/profile/profile_utils.dart
@@ -1,7 +1,0 @@
-// Package imports:
-import 'package:intl/intl.dart';
-
-String returnCurrentYear() {
-  final DateTime now = DateTime.now();
-  return DateFormat('y').format(now);
-}
