+++ lib/presentation/router/router_generator.dart
@@ -8,6 +8,8 @@ import 'package:afya_moja_core/community_list_item.dart';
 import 'package:myafyahub/presentation/health_diary/widgets/mood_selection/mood_feedback_page.dart';
 import 'package:myafyahub/presentation/home/pages/home_page.dart';
 import 'package:myafyahub/presentation/onboarding/set_security_questions/security_questions_page.dart';
+import 'package:myafyahub/presentation/profile/faqs/faq_detail_view_page.dart';
+import 'package:myafyahub/presentation/profile/faqs/profile_faqs_page.dart';
 import 'package:user_profile/term_and_conditions.dart' as webview;
 
 // Project imports:
@@ -26,23 +28,21 @@ import 'package:myafyahub/presentation/my_health/pages/my_health_diary_page.dart
 import 'package:myafyahub/presentation/my_health/pages/my_health_page.dart';
 import 'package:myafyahub/presentation/notifications/notifications_page.dart';
 import 'package:myafyahub/presentation/onboarding/set_nickname/pages/congratulations_page.dart';
-import 'package:myafyahub/presentation/onboarding/login/pages/create_new_pin_page.dart';
-import 'package:myafyahub/presentation/onboarding/login/pages/forgot_pin_page.dart';
+import 'package:myafyahub/presentation/onboarding/set_new_pin/pages/create_new_pin_page.dart';
+import 'package:myafyahub/presentation/onboarding/forgot_pin/pages/forgot_pin_page.dart';
 import 'package:myafyahub/presentation/onboarding/login/pages/login_page.dart';
 import 'package:myafyahub/presentation/onboarding/login/pages/pin_input_page.dart';
 import 'package:myafyahub/presentation/onboarding/verify_phone/pages/verify_phone_page.dart';
 import 'package:myafyahub/presentation/onboarding/set_security_questions/answer_security_question_page.dart';
 import 'package:myafyahub/presentation/onboarding/terms/terms_and_conditions_page.dart';
-import 'package:myafyahub/presentation/profile/pages/clinic_information_page.dart';
-import 'package:myafyahub/presentation/profile/pages/consent_page.dart';
+import 'package:myafyahub/presentation/profile/clinic_information/clinic_information_page.dart';
+import 'package:myafyahub/presentation/profile/consent/consent_page.dart';
 import 'package:myafyahub/presentation/profile/pages/edit_information_page.dart';
-import 'package:myafyahub/presentation/profile/pages/feedback_page.dart';
-import 'package:myafyahub/presentation/profile/pages/medical_data_page.dart';
-import 'package:myafyahub/presentation/profile/pages/personal_information_page.dart';
-import 'package:myafyahub/presentation/profile/pages/profile_faqs/faq_detail_view_page.dart';
-import 'package:myafyahub/presentation/profile/pages/profile_faqs/profile_faqs_page.dart';
-import 'package:myafyahub/presentation/profile/pages/saved_posts.dart';
-import 'package:myafyahub/presentation/profile/pages/settings_page.dart';
+import 'package:myafyahub/presentation/profile/feedback/feedback_page.dart';
+import 'package:myafyahub/presentation/profile/medical_data/medical_data_page.dart';
+import 'package:myafyahub/presentation/profile/personal_information/personal_information_page.dart';
+import 'package:myafyahub/presentation/profile/saved_posts/saved_posts.dart';
+import 'package:myafyahub/presentation/profile/settings/settings_page.dart';
 import 'package:myafyahub/presentation/profile/pages/user_profile_page.dart';
 import 'package:myafyahub/presentation/router/routes.dart';
 
@@ -219,9 +219,7 @@ class RouteGenerator {
 
       case BWRoutes.faqDetailViewPage:
         return MaterialPageRoute<FAQDetailViewPage>(
-          builder: (_) => FAQDetailViewPage(
-            payload: args as FAQContent,
-          ),
+          builder: (_) => FAQDetailViewPage(payload: args as FAQContent),
         );
 
       case BWRoutes.forgotPINPage:
