+++ test/widgets_tests/presentation/feed/widgets/suggested_group_card_test.dart
@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 // Package imports:
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail_image_network/mocktail_image_network.dart';
-import 'package:myafyahub/presentation/communities/chat_screen/widgets/suggested_group_card.dart';
+import 'package:myafyahub/presentation/communities/suggested_groups/suggested_group_card.dart';
 
 // Project imports:
 import '../../../../mocks.dart';
