+++ test/widgets_tests/presentation/onboarding/login/widgets/verify_otp_widget_test.dart
@@ -19,7 +19,7 @@ import 'package:myafyahub/domain/core/value_objects/app_strings.dart';
 import 'package:myafyahub/domain/core/value_objects/app_widget_keys.dart';
 import 'package:myafyahub/infrastructure/endpoints.dart';
 import 'package:myafyahub/presentation/core/widgets/pin_input_field_widget.dart';
-import 'package:myafyahub/presentation/onboarding/login/widgets/verify_otp_widget.dart';
+import 'package:myafyahub/presentation/onboarding/verify_phone/widgets/verify_otp_widget.dart';
 import '../../../../../mocks.dart';
 import '../../../../../test_helpers.dart';
