+++ test/widgets_tests/presentation/profile/clinic_information_page_test.dart
@@ -6,8 +6,8 @@ import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
 import 'package:myafyahub/domain/core/value_objects/app_widget_keys.dart';
-import 'package:myafyahub/presentation/profile/pages/clinic_information_page.dart';
-import 'package:myafyahub/presentation/profile/widgets/clinic_information_item_widget.dart';
+import 'package:myafyahub/presentation/profile/clinic_information/clinic_information_page.dart';
+import 'package:myafyahub/presentation/profile/clinic_information/clinic_information_item_widget.dart';
 
 void main() {
   testWidgets('ClinicInformationPage renders correctly',
