+++ test/widgets_tests/presentation/profile/pages/consent_page_test.dart
@@ -14,7 +14,7 @@ import 'package:myafyahub/application/redux/states/app_state.dart';
 import 'package:myafyahub/domain/core/value_objects/app_strings.dart';
 import 'package:myafyahub/domain/core/value_objects/app_widget_keys.dart';
 import 'package:myafyahub/presentation/onboarding/terms/terms_and_conditions_page.dart';
-import 'package:myafyahub/presentation/profile/pages/consent_page.dart';
+import 'package:myafyahub/presentation/profile/consent/consent_page.dart';
 import '../../../../mocks.dart';
 import '../../../../test_helpers.dart';
