+++ test/widgets_tests/presentation/profile/settings_page_test.dart
@@ -6,7 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:myafyahub/application/redux/states/app_state.dart';
 import 'package:myafyahub/domain/core/value_objects/app_strings.dart';
 import 'package:myafyahub/presentation/core/widgets/app_bar/custom_app_bar.dart';
-import 'package:myafyahub/presentation/profile/pages/settings_page.dart';
+import 'package:myafyahub/presentation/profile/settings/settings_page.dart';
 import '../../../test_helpers.dart';
 
 void main() {
