+++ test/widgets_tests/presentation/profile/widgets/my_health_timeline_test.dart
@@ -5,8 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 // Project imports:
-import 'package:myafyahub/presentation/profile/widgets/my_health_timeline.dart';
-import 'package:myafyahub/presentation/profile/widgets/timeline_indicator.dart';
+import 'package:myafyahub/presentation/profile/health_timeline/my_health_timeline.dart';
+import 'package:myafyahub/presentation/profile/health_timeline/timeline_indicator.dart';
 
 void main() {
   group('MyHealthTimeline', () {
